Feat: Add lint

+++ contracts/Request.sol
@@ -7,17 +7,14 @@ struct RequestTrace {
 
 struct Request {
     address userAddress;
-    uint256 nonce;
+    uint nonce;
     bytes32 serviceType;
     bytes signature;
-    uint256 createdAt;
+    uint createdAt;
 }
 
 library RequestLibrary {
-    function verify(
-        Request memory request,
-        address serviceProviderAddress
-    ) internal pure returns (bool) {
+    function verify(Request memory request, address serviceProviderAddress) internal pure returns (bool) {
         bytes32 message = prefixed(
             keccak256(
                 abi.encodePacked(
@@ -33,9 +30,7 @@ library RequestLibrary {
         return recoverSigner(message, request.signature) == request.userAddress;
     }
 
-    function splitSignature(
-        bytes memory sig
-    ) internal pure returns (uint8 v, bytes32 r, bytes32 s) {
+    function splitSignature(bytes memory sig) internal pure returns (uint8 v, bytes32 r, bytes32 s) {
         require(sig.length == 65);
 
         assembly {
@@ -50,18 +45,12 @@ library RequestLibrary {
         return (v, r, s);
     }
 
-    function recoverSigner(
-        bytes32 message,
-        bytes memory sig
-    ) internal pure returns (address) {
+    function recoverSigner(bytes32 message, bytes memory sig) internal pure returns (address) {
         (uint8 v, bytes32 r, bytes32 s) = splitSignature(sig);
         return ecrecover(message, v, r, s);
     }
 
     function prefixed(bytes32 hash) internal pure returns (bytes32) {
-        return
-            keccak256(
-                abi.encodePacked("\x19Ethereum Signed Message:\n32", hash)
-            );
+        return keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", hash));
     }
 }
